# schemas/openapi/api-v1.yaml
openapi: 3.1.0
info:
  title: RaptorFlow API
  version: 0.1.0
servers:
  - url: https://api.raptorflow.example.com
paths:
...
  /api/v1/campaigns/{id}:
    get:
      summary: Get campaign with moves and tasks
      responses:
        "200":
          description: Campaign detail
        "404":
          description: Campaign not found
  /api/v1/campaigns/{id}/status:
    patch:
      summary: Update campaign status
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [draft, active, paused, completed, archived]
      responses:
        "200":
          description: Status updated
        "400":
          description: Invalid status
        "404":
          description: Campaign not found
  /api/v1/campaigns/{id}/moves:
    get:
      summary: List campaign moves
      responses:
        "200":
          description: Move list
        "404":
          description: Campaign not found
    post:
      summary: Create campaign move
      responses:
        "201":
          description: Move created
        "404":
          description: Campaign not found
  /api/v1/campaigns/{id}/moves/{move_id}/status:
    patch:
      summary: Update move status
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [planned, in_progress, completed, skipped]
      responses:
        "200":
          description: Status updated
        "400":
          description: Invalid status
  /api/v1/campaigns/{id}/tasks:
    get:
      summary: List campaign tasks
      responses:
        "200":
          description: Task list
        "404":
          description: Campaign not found
    post:
      summary: Create campaign task
      responses:
        "201":
          description: Task created
        "404":
          description: Campaign not found
  /api/v1/campaigns/{id}/tasks/{task_id}/status:
    patch:
      summary: Update task status
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [pending, in_progress, completed, cancelled]
      responses:
        "200":
          description: Status updated
        "400":
          description: Invalid status
  /api/v1/campaigns/{id}/brief:
    get:
      summary: Get campaign brief
      responses:
        "200":
          description: Brief found or null
        "404":
          description: Campaign not found
    post:
      summary: Create campaign brief
      responses:
        "201":
          description: Brief created
        "404":
          description: Campaign not found
  /api/v1/campaigns/{id}/brief/status:
    patch:
      summary: Update brief status
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [draft, submitted, approved, rejected]
      responses:
        "200":
          description: Status updated
        "400":
          description: Invalid status
        "404":
          description: Brief not found
...